# tests/contract/oac_irs/irs-baseline/v1/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentRegistryEnabledRequest": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "AgentRegistryEnabledRequest",
        "type": "object"
      },
...
      "AgentRegistryItem": {
        "additionalProperties": false,
        "properties": {
          "agent_id": {
            "minLength": 1,
            "title": "Agent Id",
            "type": "string"
          },
          "allowed_user_tags": {
            "items": {
              "type": "string"
            },
            "title": "Allowed User Tags",
            "type": "array"
          },
          "bot_id": {
            "default": "",
            "title": "Bot Id",
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "negative_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Negative Keywords",
            "type": "array"
          },
          "positive_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Positive Keywords",
            "type": "array"
          },
          "route_path": {
            "default": "",
            "title": "Route Path",
            "type": "string"
          }
        },
        "required": [
          "agent_id",
          "name",
          "description"
        ],
        "title": "AgentRegistryItem",
        "type": "object"
      },
      "AgentRegistryUpsertRequest": {
        "additionalProperties": false,
        "properties": {
          "agent_id": {
            "minLength": 1,
            "title": "Agent Id",
            "type": "string"
          },
          "allowed_user_tags": {
            "items": {
              "type": "string"
            },
            "title": "Allowed User Tags",
            "type": "array"
          },
          "bot_id": {
            "default": "",
            "title": "Bot Id",
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "negative_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Negative Keywords",
            "type": "array"
          },
          "positive_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Positive Keywords",
            "type": "array"
          },
          "route_path": {
            "default": "",
            "title": "Route Path",
            "type": "string"
          }
        },
        "required": [
          "agent_id",
          "name",
          "description"
        ],
        "title": "AgentRegistryUpsertRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "OAC Central Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/agent-registry": {
      "get": {
        "operationId": "list_agent_registry_api_v1_admin_agent_registry_get",
        "parameters": [
          {
            "in": "query",
            "name": "enabled_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Enabled Only",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "x-admin-sync-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Sync-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentRegistryItem"
                  },
                  "title": "Response List Agent Registry Api V1 Admin Agent Registry Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Agent Registry",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "upsert_agent_registry_item_api_v1_admin_agent_registry_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-sync-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Sync-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRegistryUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRegistryItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Agent Registry Item",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/admin/agent-registry/{agent_id}": {
      "delete": {
        "operationId": "delete_agent_registry_item_api_v1_admin_agent_registry__agent_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-sync-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Sync-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Agent Registry Item",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "update_agent_registry_item_api_v1_admin_agent_registry__agent_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-sync-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Sync-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRegistryUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRegistryItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Agent Registry Item",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/agent-registry/{agent_id}/enabled": {
      "patch": {
        "operationId": "set_agent_registry_item_enabled_api_v1_admin_agent_registry__agent_id__enabled_patch",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-sync-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Sync-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRegistryEnabledRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRegistryItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Agent Registry Item Enabled",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}